Big change the the URL mappings.
Sitemesh decorators never really worked perfectly for us, see:
http://forums.opensymphony.com/message.jspa?messageID=349400#349400
Problem would be that getServletPath() was just "/site" and we couldn't decorate based on that.

we web.xml mapped dispatcher to "/site/*" bc that let "/com.apress.gwt/*" URL's NOT got to dispatcher servlet, which is what we wanted. 

Problems with this were 
1) all URL's needed the annoying /site/ prefix
2) these sitemesh issues.

This reve re-does everything w/ dispatcher mapped to *.html. This break GWT, so we created a FileServlet that we map to the GWT dirs which will override the dispatcher mappings.